refactor(Footer): update import and delete old image
