+++ src/layouts/AppLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import AppBar from "../components/AppBar/AppBar";
-import Footer from "../components/Footer";
+import Footer from "../components/Footer/Footer";
 import CartCleaner from "../features/cart/components/CartCleaner";
 
 export default function AppLayout() {
